add sweetAlert on delete

+++ src/pages/examples/ListEvents.js
@@ -28,6 +28,7 @@ import { GeneralInfoForm } from "../../components/Forms";
 import EventService from "../../services/event.services";
 import Profile3 from "../../assets/img/team/profile-picture-3.jpg";
 import { Fragment } from "react";
+import Swal from "sweetalert2";
 
 export default (props) => {
   const [events, setEvents] = useState([]);
@@ -46,17 +47,25 @@ export default (props) => {
       });
   };
   const handleDelete = (id) => {
-    console.log(id);
-    EventService.deleteEvent(id)
-      .then((response) => {
-        console.log(response);
-        toast.success(response.data.message);
-        refreshList();
-      })
-      .catch((error) => {
-        console.log(error);
-        toast.error("Internal server error");
-      });
+    Swal.fire({
+      title: "Do you want to delete the Event?",
+      showCancelButton: true,
+      confirmButtonText: `Delete`,
+    }).then((result) => {
+      if (result.isConfirmed) {
+        EventService.deleteEvent(id)
+          .then((response) => {
+            //console.log(response);
+            Swal.fire("Event Deleted!", "", "success");
+            setEvents(events);
+            refreshList();
+          })
+          .catch((error) => {
+            console.log(error);
+            toast.error("Internal server error");
+          });
+      }
+    });
   };
 
   const handleEdit = (id) => {
